task_2/rerequirements.txt
opencv-python==4.5.3.56
numpy==1.21.2
pandas==1.3.2
seaborn==0.11.2
Pillow==8.3.2
matplotlib==3.4.3
tqdm==4.62.2

